requirements.txt
# LiveKit and Core
livekit==0.15.1
livekit-agents>=0.8.2
livekit-plugins-openai>=0.9.0

# AI and ML
openai>=1.12.0
google-generativeai>=0.5.0
anthropic>=0.21.0
ollama>=0.1.7
whisper>=1.1.10

# Computer Vision
opencv-python>=4.9.0
Pillow>=11.2.1
pytesseract>=0.3.10
numpy>=1.26.0
screeninfo>=0.8.1

# Remote Control
pyautogui>=0.9.54
pynput>=1.7.6

# Audio Processing
pydub>=0.25.1
sounddevice>=0.4.6
soundfile>=0.12.1

# Web and API
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
requests>=2.31.0

# Database and Storage
redis>=5.0.0
pydantic>=2.6.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
colorlog>=6.8.0
psutil>=5.9.8